+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 7.2
 - Add link to Student Info in AddDrop.php
 - Limit students to User schools in AddDrop.php
 - Order Day, Month & Year inputs depending on User date preference in Date.php
+- SQL fix only display enrolled students in AddStudents.php
 
 Changes in 7.1.2
 ----------------